code/bots/graphql_api_gateway.py
#!/usr/bin/env python3
"""
GraphQL API Gateway v13.0
Unified GraphQL API with federation, real-time subscriptions
Apollo Federation for microservices
"""

import os
import sys
import asyncio
import logging
from datetime import datetime
from typing import Dict, List, Optional, Any
from dataclasses import dataclass
import strawberry
from strawberry.asgi import GraphQL
from strawberry.subscriptions import GRAPHQL_TRANSPORT_WS_PROTOCOL
import uvicorn
from fastapi import FastAPI, WebSocket
from fastapi.middleware.cors import CORSMiddleware

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

################################################################################
# GraphQL Types
################################################################################

@strawberry.type
class Deployment:
    """Deployment type"""
    id: str
    name: str
    namespace: str
    replicas: int
    available_replicas: int
    status: str
    created_at: str

@strawberry.type
class Metric:
    """Metric type"""
    name: str
    value: float
    unit: str
    timestamp: str
    labels: strawberry.scalars.JSON

@strawberry.type
class SecurityThreat:
    """Security threat type"""
    threat_id: str
    threat_type: str
    severity: str
    source_ip: str
    description: str
    mitigated: bool
    timestamp: str

@strawberry.type
class CostForecast:
    """Cost forecast type"""
    forecast_id: str
    provider: str
    current_cost: float
    forecasted_cost: float
    trend: str
    confidence: float

@strawberry.type
class Event:
    """Real-time event type"""
    event_id: str
    event_type: str
    source: str
    payload: strawberry.scalars.JSON
    timestamp: str

################################################################################
# GraphQL Queries
################################################################################

@strawberry.type
class Query:
    """Root query"""
    
    @strawberry.field
    async def deployments(self, namespace: Optional[str] = None) -> List[Deployment]:
        """Get all deployments"""
        # Mock data - integrate with real deployment engine
        return [
            Deployment(
                id="deploy-1",
                name="web-app",
                namespace=namespace or "default",
                replicas=3,
                available_replicas=3,
                status="Running",
                created_at=datetime.now().isoformat()
            )
        ]
    
    @strawberry.field
    async def deployment(self, id: str) -> Optional[Deployment]:
        """Get deployment by ID"""
        return Deployment(
            id=id,
            name="web-app",
            namespace="default",
            replicas=3,
            available_replicas=3,
            status="Running",
            created_at=datetime.now().isoformat()
        )
    
    @strawberry.field
    async def metrics(self, metric_names: List[str]) -> List[Metric]:
        """Get metrics"""
        return [
            Metric(
                name=name,
                value=85.5,
                unit="percent",
                timestamp=datetime.now().isoformat(),
                labels={"component": "cpu"}
            )
            for name in metric_names
        ]
    
    @strawberry.field
    async def security_threats(self, severity: Optional[str] = None) -> List[SecurityThreat]:
        """Get security threats"""
        return [
            SecurityThreat(
                threat_id="threat-1",
                threat_type="intrusion",
                severity=severity or "high",
                source_ip="192.168.1.100",
                description="Suspicious activity detected",
                mitigated=False,
                timestamp=datetime.now().isoformat()
            )
        ]
    
    @strawberry.field
    async def cost_forecast(self, provider: str) -> CostForecast:
        """Get cost forecast"""
        return CostForecast(
            forecast_id="forecast-1",
            provider=provider,
            current_cost=1250.50,
            forecasted_cost=1450.75,
            trend="increasing",
            confidence=0.87
        )

################################################################################
# GraphQL Mutations
################################################################################

@strawberry.type
class Mutation:
    """Root mutation"""
    
    @strawberry.mutation
    async def scale_deployment(self, deployment_id: str, replicas: int) -> Deployment:
        """Scale deployment"""
        logger.info(f"Scaling deployment {deployment_id} to {replicas} replicas")
        
        return Deployment(
            id=deployment_id,
            name="web-app",
            namespace="default",
            replicas=replicas,
            available_replicas=replicas,
            status="Scaling",
            created_at=datetime.now().isoformat()
        )
    
    @strawberry.mutation
    async def mitigate_threat(self, threat_id: str) -> SecurityThreat:
        """Mitigate security threat"""
        logger.info(f"Mitigating threat {threat_id}")
        
        return SecurityThreat(
            threat_id=threat_id,
            threat_type="intrusion",
            severity="high",
            source_ip="192.168.1.100",
            description="Threat mitigated",
            mitigated=True,
            timestamp=datetime.now().isoformat()
        )

################################################################################
# GraphQL Subscriptions (Real-Time)
################################################################################

@strawberry.type
class Subscription:
    """Root subscription for real-time updates"""
    
    @strawberry.subscription
    async def events(self, event_types: Optional[List[str]] = None) -> Event:
        """Subscribe to real-time events"""
        while True:
            await asyncio.sleep(5)  # Simulate event generation
            
            yield Event(
                event_id="event-123",
                event_type="deployment",
                source="kubernetes",
                payload={"message": "Deployment updated"},
                timestamp=datetime.now().isoformat()
            )
    
    @strawberry.subscription
    async def metrics_stream(self, metric_names: List[str]) -> Metric:
        """Subscribe to real-time metrics"""
        import random
        
        while True:
            await asyncio.sleep(2)
            
            for name in metric_names:
                yield Metric(
                    name=name,
                    value=random.uniform(50, 100),
                    unit="percent",
                    timestamp=datetime.now().isoformat(),
                    labels={"component": name}
                )
    
    @strawberry.subscription
    async def security_alerts(self) -> SecurityThreat:
        """Subscribe to security alerts"""
        while True:
            await asyncio.sleep(10)
            
            yield SecurityThreat(
                threat_id=f"threat-{datetime.now().timestamp()}",
                threat_type="intrusion",
                severity="high",
                source_ip="192.168.1.100",
                description="Real-time security alert",
                mitigated=False,
                timestamp=datetime.now().isoformat()
            )

################################################################################
# GraphQL Schema
################################################################################

schema = strawberry.Schema(
    query=Query,
    mutation=Mutation,
    subscription=Subscription
)

################################################################################
# FastAPI Application
################################################################################

app = FastAPI(
    title="GraphQL API Gateway v13.0",
    description="Unified GraphQL API for all services",
    version="13.0.0"
)

# CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# GraphQL endpoint
graphql_app = GraphQL(
    schema,
    subscription_protocols=[GRAPHQL_TRANSPORT_WS_PROTOCOL]
)

app.add_route("/graphql", graphql_app)
app.add_websocket_route("/graphql", graphql_app)

@app.get("/")
async def root():
    """Root endpoint"""
    return {
        "service": "GraphQL API Gateway v13.0",
        "endpoints": {
            "graphql": "/graphql",
            "playground": "/graphql (GET)"
        }
    }

@app.get("/health")
async def health():
    """Health check"""
    return {"status": "healthy"}

################################################################################
# CLI
################################################################################

def main():
    """Main entry point"""
    logger.info("🚀 Starting GraphQL API Gateway v13.0")
    
    port = int(os.getenv('GRAPHQL_PORT', '8000'))
    
    uvicorn.run(
        app,
        host="0.0.0.0",
        port=port,
        log_level="info"
    )

if __name__ == '__main__':
    main()
